Ref: move logics to services move route functions to controllers move menuma logic to MenumaController
